requirements.txt
lightkurve==2.4.2
astroquery==0.4.7
astropy==6.0.1
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.0
scikit-learn==1.5.0
scipy==1.13.1
torch==2.3.0
torchvision==0.18.0
wandb==0.17.0
pyyaml==6.0.1
tqdm==4.66.4
requests==2.32.3
seaborn==0.13.2
imbalanced-learn==0.12.3
jupyterlab==4.2.1
pytest==8.2.0
ruff==0.4.4
mypy==1.10.0
psutil==5.9.8
boto3==1.34.110
botocore<1.42.85